long numbers to string fixed

+++ chr.h
@@ -34,11 +34,11 @@ void setAscii(char *a, int length = -1) {
 	for (int x = 0; x < size; x++) s[x] = (char)(*a++);
 }
 
-int intToStr(int i, char *s) {
+long intToStr(long i, char *s) {
 	char *c = s;
-	int n = 0;
+	long n = 0;
 	while (true) {
-		unsigned int i10 = i / 10;
+		unsigned long i10 = i / 10;
 		*s = (i - (i10 * 10)) + '0';
 		i = i10;
 		s++;
@@ -56,11 +56,11 @@ int intToStr(int i, char *s) {
 
 void dblToStr (double d, char *s) {
 	if (d < 0) d = -d, s[0] = '-', s++;
-	int n = d;
-	int w = intToStr(n, s);
+	long n = d;
+	long w = intToStr(n, s);
 	s += w;
 	d = d - n;
-	int q = d * 100000;
+	long q = d * 100000;
 	d *= 100000000;
 	{ // fix
 		q = d - q * 1000;
